Fixed process facility programName() returning an empty string
This seemed to have not been implemented. Implementing this required that the
cmdLine variable from main() be promoted to global scope, because that is what
contained the bytecode name passed on the command line. This in turn required
splitting CommandLineArguments constructor to a default constructor and an
init() method. Fortunately all of this could be done in a quite straightforward
manner.